requirements.txt
biopython==1.81
easydict==1.11
PyYAML==6.0
numpy==1.25.1
pandas==2.0.3
scrikit-learn==1.3.0
scipy==1.11.1
tensorboard==2.13.0
tqdm==4.65.0
warmup-scheduler==0.3
